\name{ConvertSoCalFiles}
\alias{ConvertSoCalFiles}
\alias{ConvertSFBayFiles}
\title{
Convert Reference and Example User Files
}
\description{
Converts text files to .Rdata format for use in package. Input files are in inst/extdata. During package development, these fuctions are used to make the .Rdata files, which are manually moved to the data directory.
}
\usage{
ConvertSoCalFiles()
ConvertSFBayFiles()
}
