\name{socal.example.taxa}
\alias{socal.example.taxa}
\alias{socal.example.habitat}
\alias{sfbay.example.taxa}
\alias{sfbay.example.habitat}
\docType{data}
\title{
Example User Data Sets. 
}
\description{
Example data sets for the Southern California Bight and San Francisco Bay. Used for testing.
}
\details{Southern California Bight:
\code{socal.example.taxa},
\code{socal.example.habitat}.

San Francisco Bay:
\code{sfbay.example.taxa},
\code{sfbay.example.habitat}.
}

\keyword{datasets}
