\name{socal.reference.taxa}
\alias{socal.reference.taxa}
\alias{socal.reference.groups}
\alias{socal.reference.group.means}
\alias{socal.reference.covariance}
\alias{socal.reference.taxa}
\alias{socal.reference.taxa}

\alias{sfbay.reference.taxa}
\alias{sfbay.reference.groups}
\alias{sfbay.reference.group.means}
\alias{sfbay.reference.covariance}
\alias{sfbay.reference.taxa}
\alias{sfbay.reference.taxa}

\docType{data}
\title{
Reference Data Sets 
}
\description{
Reference Data Sets for the Southern California Bight and San Francisco Bay. Provided by ecologists.
}
\details{Southern California Bight:
\code{socal.reference.taxa},
\code{socal.reference.groups},
\code{socal.reference.group.means},
\code{socal.reference.covariance},
\code{socal.reference.taxa},
\code{socal.reference.taxa}.

San Francisco Bay:
\code{sfbay.reference.taxa},
\code{sfbay.reference.groups},
\code{sfbay.reference.group.means},
\code{sfbay.reference.covariance},
\code{sfbay.reference.taxa},
\code{sfbay.reference.taxa}.
}

\keyword{datasets}
